+++ avatars/integration_test/run_complete.py
@@ -6,7 +6,7 @@ from avatars.models import (
     ColumnType,
     DatasetResponse,
     JobCreate,
-    JobParameters,
+    AvatarizationParameters,
     PatchDataset,
 )
 
@@ -54,7 +54,7 @@ def main():
     print(response, "\n")
 
     # All calls to Jobs
-    job_create = JobCreate(dataset_id=dataset_id, parameters=JobParameters(k=20))
+    job_create = JobCreate(dataset_id=dataset_id, parameters=AvatarizationParameters(k=20))
     response = client.jobs.create_job(request=job_create)
     print(response, "\n")
     job_id = response.id