+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## 0.1.16
+- Add forgotten password endpoint
+- Add reset password endpoint
+- JobParameters becomes AvatarizationParameters
+- Add DCR and NNDR to privacy metrics
+
 ## 0.1.15
 
 - Handle category dtype
-- Fix download of reports
 - Fix dtype casting of datetime columns
 - Add ability to login with email
 - Add filtering options to `find_users`